Update and fix many minor PHPCS issues
This codebase uses a very different style. It's very carefully crafted,
just different. I really do not want to destroy any of this. So all I
do in this patch is to fix issues that are actual mistakes. But I won't
enforce the MediaWiki style guide on this codebase, and don't suggest
anyone else to do this either.

Change-Id: Iddb1599b5e3d993539ca14c0c5f1a9c1d8c2908c