Fix documentation and some problematic isset() usages
isset() does two things: It doesn't only check if a variable
contains null. It also checks if a variable exists. This is helpful
when dealing with non-existing array elements, for example. But in
these cases here we know the variable must exist.

The problem is that isset() hides programming errors. For example:

 function ( $example ) {
     if ( isset( $examlpe ) ) …
 }

Nothing would ever happen here even if there is a typo. No error
from PHP and no error from the IDE (maybe a warning if the IDE is
clever). We can avoid this when we use isset() only in situations
where we want this behavior.

Change-Id: I98c3fe5bd48fa9c4356ad4a7a882bb4e94bf3c0d