Replace deprecated $wgHooks access with HookContainer
This is the most minimal change I can think of to get rid of the
deprecation warning. A better solution is to use the hook handler
interfaces. I will do this in a separate patch after this one.

Verified via Special:Version. The parser tags and functions still
appear there.

Change-Id: I9b59bd4682fcceab1c2905605d87a8b607c8a3e5